backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
bitarray==2.9.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.7
cytoolz==0.12.3
daphne==4.1.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-keycloak-admin==0.0.4
django-macaddress==1.8.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
dnspython==2.6.1
-e git+https://github.com/ErikPolzin/drf-dynamic-fields@f2f285a3402d28055a71266277673739e79e4533#egg=drf_dynamic_fields
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==0.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.3.7
lru-dict==1.2.0
msgpack==1.0.8
multidict==6.0.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
netaddr==1.3.0
packaging==24.0
parsimonious==0.10.0
prompt_toolkit==3.0.47
protobuf==5.27.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pymongo==4.7.3
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keycloak==4.0.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
rpds-py==0.18.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
toolz==0.12.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
web3==6.19.0
websockets==12.0
yarl==1.9.4
zope.interface==6.4.post2
